move files and update name
