+++ src/routes/index.ts
@@ -3,7 +3,8 @@ import swaggerUi from 'swagger-ui-express';
 import swaggerDocument from '../api-docs/swagger-output.json';
 import path from 'path';
 import userRouter from './user';
-import authRouter from './auth'
+import authRouter from './auth';
+import homeRouter from './home';
 import taskRouter from './task';
 import categoryRouter from './category';
 import logRouter from './log';
@@ -24,7 +25,8 @@ router.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 // });
 
 router.use('/user', userRouter);
-router.use('/auth', authRouter)
+router.use('/auth', authRouter);
+router.use('/home', homeRouter);
 router.use('/task', taskRouter);
 router.use('/category', categoryRouter);
 router.use('/log', logRouter);
