+++ src/views/home.ejs
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <title>Weather</title>
+    <title>Home</title>
     <link rel="stylesheet" href="/styles/home.css">
     <script defer src="/scripts/home.js"></script>
 </head>
